refactor(core): lift MeshAgent + configFromEnv into core for shared adapters
