+++ packages/core/src/index.ts
@@ -4,3 +4,5 @@ export * from "./endpoint.js";
 export * from "./registry.js";
 export * from "./connector.js";
 export * from "./runtime.js";
+export * from "./config.js";
+export * from "./mesh.js";
